Used ELF symbol table to find function names not present in DWARF sections
Before this commit, when the backtrace implementation could not detect a
function name in DWARF symbols, it would print "Name unknown". Now,
similar to libunwind, the crate uses the ELF symbol table to locate
functions at the given address when there is no information about the
name in the DWARF sections.